lara/control: port Lara_Control to TRX
This merges Lara's primary control routine into TRX and splits it
out with some additional helper functions to improve readability.
