+++ src/tr1/game/lara/common.c
@@ -17,15 +17,10 @@
 
 #include <libtrx/config.h>
 #include <libtrx/game/camera.h>
-#include <libtrx/game/collision.h>
 #include <libtrx/game/lara.h>
 #include <libtrx/game/math.h>
-#include <libtrx/game/music.h>
 #include <libtrx/log.h>
 #include <libtrx/utils.h>
-
-#define LARA_MOVE_TIMEOUT 90
-#define LARA_UW_DAMAGE 5
 
 LARA_INFO *Lara_GetLaraInfo(void)
 {
@@ -35,287 +30,6 @@ LARA_INFO *Lara_GetLaraInfo(void)
 ITEM *Lara_GetItem(void)
 {
     return g_LaraItem;
-}
-
-void Lara_Control(void)
-{
-    COLL_INFO coll = {};
-
-    ITEM *item = g_LaraItem;
-    const ROOM *const room = Room_Get(item->room_num);
-    const bool room_submerged = (room->flags & RF_UNDERWATER) != 0;
-    const int32_t water_depth = Lara_GetWaterDepth(
-        item->pos.x, item->pos.y, item->pos.z, item->room_num);
-    const int32_t water_height = Room_GetWaterHeight(
-        item->pos.x, item->pos.y, item->pos.z, item->room_num);
-    const int32_t water_height_diff =
-        water_height == NO_HEIGHT ? NO_HEIGHT : item->pos.y - water_height;
-    g_Lara.water_surface_dist = -water_height_diff;
-
-    if (g_Lara.interact_target.is_moving
-        && g_Lara.interact_target.move_count++ > LARA_MOVE_TIMEOUT) {
-        g_Lara.interact_target.is_moving = false;
-        g_Lara.gun_status = LGS_ARMLESS;
-    }
-
-    if (item->hit_points > 0 && g_Config.debug.enable_invulnerability) {
-        item->hit_points = LARA_MAX_HITPOINTS;
-    }
-
-    if (g_InputDB.level_skip_cheat) {
-        Lara_Cheat_EndLevel();
-    }
-
-    if (g_InputDB.item_cheat) {
-        Lara_Cheat_GiveAllItems();
-    }
-
-    if (g_Lara.water_status != LWS_CHEAT && g_InputDB.fly_cheat) {
-        Lara_Cheat_EnterFlyMode();
-    }
-
-    switch (g_Lara.water_status) {
-    case LWS_ABOVE_WATER: {
-        if (g_Config.gameplay.enable_wading
-            && (water_height_diff == NO_HEIGHT
-                || water_height_diff < LARA_WADE_DEPTH)) {
-            break;
-        }
-
-        if (g_Config.gameplay.enable_wading
-            && water_depth <= LARA_SWIM_DEPTH - STEP_L) {
-            if (water_height_diff > LARA_WADE_DEPTH) {
-                g_Lara.water_status = LWS_WADE;
-                if (!item->gravity) {
-                    item->goal_anim_state = LS_STOP;
-                }
-            }
-        } else if (room_submerged) {
-            g_Lara.water_status = LWS_UNDERWATER;
-            g_Lara.air = LARA_MAX_AIR;
-            item->pos.y += 100;
-            item->gravity = false;
-            Lara_UpdateRoomToHeight(0);
-            Sound_StopEffect(SFX_LARA_FALL);
-            if (item->current_anim_state == LS_SWAN_DIVE) {
-                item->goal_anim_state = LS_DIVE;
-                item->rot.x = -45 * DEG_1;
-                Lara_Animate(item);
-                item->fall_speed *= 2;
-            } else if (item->current_anim_state == LS_FAST_DIVE) {
-                item->goal_anim_state = LS_DIVE;
-                item->rot.x = -85 * DEG_1;
-                Lara_Animate(item);
-                item->fall_speed *= 2;
-            } else {
-                item->current_anim_state = LS_DIVE;
-                item->goal_anim_state = LS_SWIM;
-                Item_SwitchToAnim(item, LA_FREEFALL_TO_UNDERWATER, 0);
-                item->rot.x = -45 * DEG_1;
-                item->fall_speed = (item->fall_speed * 3) / 2;
-            }
-            g_Lara.head_rot.x = 0;
-            g_Lara.head_rot.y = 0;
-            g_Lara.torso_rot.x = 0;
-            g_Lara.torso_rot.y = 0;
-            Spawn_Splash(item);
-        }
-
-        break;
-    }
-
-    case LWS_UNDERWATER: {
-        if (room_submerged) {
-            break;
-        }
-
-        if (water_depth == NO_HEIGHT || ABS(water_height_diff) >= STEP_L) {
-            g_Lara.water_status = LWS_ABOVE_WATER;
-            g_Lara.gun_status = LGS_ARMLESS;
-            item->current_anim_state = LS_JUMP_FORWARD;
-            item->goal_anim_state = LS_JUMP_FORWARD;
-            Item_SwitchToAnim(item, LA_FALL_START, 0);
-            item->speed = item->fall_speed / 4;
-            item->fall_speed = 0;
-            item->gravity = true;
-            item->rot.x = 0;
-            item->rot.z = 0;
-            g_Lara.head_rot.x = 0;
-            g_Lara.head_rot.y = 0;
-            g_Lara.torso_rot.x = 0;
-            g_Lara.torso_rot.y = 0;
-        } else {
-            g_Lara.water_status = LWS_SURFACE;
-            g_Lara.dive_timer = LARA_DIVE_WAIT + 1;
-            item->current_anim_state = LS_SURF_TREAD;
-            item->goal_anim_state = LS_SURF_TREAD;
-            Item_SwitchToAnim(item, LA_UNDERWATER_TO_ONWATER, 0);
-            item->fall_speed = 0;
-            item->pos.y += 1 - water_height_diff;
-            item->rot.x = 0;
-            item->rot.z = 0;
-            g_Lara.head_rot.x = 0;
-            g_Lara.head_rot.y = 0;
-            g_Lara.torso_rot.x = 0;
-            g_Lara.torso_rot.y = 0;
-            Lara_UpdateRoomToHeight(-LARA_HEIGHT / 2);
-            Sound_Effect(SFX_LARA_BREATH, &item->pos, SPM_ALWAYS);
-        }
-        break;
-    }
-
-    case LWS_SURFACE: {
-        if (room_submerged) {
-            break;
-        }
-
-        if (g_Config.gameplay.enable_wading
-            && water_height_diff > LARA_WADE_DEPTH) {
-            g_Lara.water_status = LWS_WADE;
-            Item_SwitchToAnim(item, LA_STAND_IDLE, 0);
-            item->current_anim_state = LS_STOP;
-            item->goal_anim_state = LS_WADE;
-            Item_Animate(item);
-            item->fall_speed = 0;
-        } else {
-            g_Lara.water_status = LWS_ABOVE_WATER;
-            g_Lara.gun_status = LGS_ARMLESS;
-            item->current_anim_state = LS_JUMP_FORWARD;
-            item->goal_anim_state = LS_JUMP_FORWARD;
-            Item_SwitchToAnim(item, LA_FALL_START, 0);
-            item->speed = item->fall_speed / 4;
-            item->fall_speed = 0;
-            item->gravity = true;
-            item->rot.x = 0;
-            item->rot.z = 0;
-            g_Lara.head_rot.x = 0;
-            g_Lara.head_rot.y = 0;
-            g_Lara.torso_rot.x = 0;
-            g_Lara.torso_rot.y = 0;
-        }
-        break;
-    }
-
-    case LWS_WADE: {
-        g_Camera.target_elevation = CAM_WADE_ELEVATION;
-
-        if (water_height_diff < LARA_WADE_DEPTH) {
-            g_Lara.water_status = LWS_ABOVE_WATER;
-            if (item->current_anim_state == LS_WADE) {
-                item->goal_anim_state = LS_RUN;
-            }
-        } else if (water_height_diff > LARA_SWIM_DEPTH) {
-            g_Lara.water_status = LWS_SURFACE;
-            item->pos.y += 1 - water_height_diff;
-
-            LARA_ANIMATION anim;
-            switch (item->current_anim_state) {
-            case LS_WALK_BACK:
-                item->goal_anim_state = LS_SURF_BACK;
-                anim = LA_ONWATER_IDLE_TO_SWIM_BACK;
-                break;
-
-            case LS_STEP_RIGHT:
-                item->goal_anim_state = LS_SURF_RIGHT;
-                anim = LA_ONWATER_SWIM_RIGHT;
-                break;
-
-            case LS_STEP_LEFT:
-                item->goal_anim_state = LS_SURF_LEFT;
-                anim = LA_ONWATER_SWIM_LEFT;
-                break;
-
-            default:
-                item->goal_anim_state = LS_SURF_SWIM;
-                anim = LA_ONWATER_SWIM_FORWARD;
-                break;
-            }
-
-            item->current_anim_state = item->goal_anim_state;
-            Item_SwitchToAnim(item, anim, 0);
-
-            item->rot.z = 0;
-            item->rot.x = 0;
-            item->gravity = false;
-            item->fall_speed = 0;
-            g_Lara.dive_timer = 0;
-            g_Lara.torso_rot.y = 0;
-            g_Lara.torso_rot.x = 0;
-            g_Lara.head_rot.y = 0;
-            g_Lara.head_rot.x = 0;
-            Lara_UpdateRoomToHeight(-LARA_HEIGHT / 2);
-        }
-        break;
-    }
-
-    default:
-        break;
-    }
-
-    if (item->hit_points <= 0) {
-        item->hit_points = -1;
-        if (!g_Lara.death_timer) {
-            Music_Stop();
-            Stats_AddDeath();
-        }
-        g_Lara.death_timer++;
-        // make sure the enemy healthbar is no longer rendered. If g_Lara later
-        // is resurrected with DOZY, she should no longer aim at the target.
-        g_Lara.target = nullptr;
-
-        if (g_LaraItem->flags & IF_ONE_SHOT) {
-            return;
-        }
-    } else if (Room_IsAbyssHeight(item->pos.y)) {
-        item->hit_points = -1;
-        g_Lara.death_timer = 9 * LOGIC_FPS;
-    }
-
-    Camera_MoveManual();
-
-    switch (g_Lara.water_status) {
-    case LWS_ABOVE_WATER:
-    case LWS_WADE:
-        g_Lara.air = LARA_MAX_AIR;
-        Lara_HandleAboveWater(item, &coll);
-        break;
-
-    case LWS_UNDERWATER:
-        if (item->hit_points >= 0) {
-            g_Lara.air--;
-            if (g_Lara.air < 0) {
-                g_Lara.air = -1;
-                Lara_TakeDamage(LARA_UW_DAMAGE, false);
-            }
-        }
-        Lara_HandleUnderwater(item, &coll);
-        break;
-
-    case LWS_SURFACE:
-        if (item->hit_points >= 0) {
-            g_Lara.air += 10;
-            if (g_Lara.air > LARA_MAX_AIR) {
-                g_Lara.air = LARA_MAX_AIR;
-            }
-        }
-        Lara_HandleSurface(item, &coll);
-        break;
-
-    case LWS_CHEAT:
-        item->hit_points = LARA_MAX_HITPOINTS;
-        g_Lara.death_timer = 0;
-        Lara_HandleUnderwater(item, &coll);
-        if (g_Input.slow && !g_Input.look && !g_Input.fly_cheat) {
-            Lara_Cheat_ExitFlyMode();
-        }
-        break;
-
-    default:
-        break;
-    }
-
-    Stats_AddDistanceTravelled(item->pos, g_Lara.last_pos);
-    g_Lara.last_pos = item->pos;
 }
 
 void Lara_SwapMeshExtra(void)
