+++ src/tr1/game/lara/common.h
@@ -9,8 +9,6 @@
 
 #include <stdint.h>
 
-void Lara_Control(void);
-
 void Lara_ControlExtra(int16_t item_num);
 
 void Lara_Initialise(const GF_LEVEL *level);
