+++ src/tr1/game/lara/control.h
@@ -5,7 +5,3 @@
 #include "global/types.h"
 
 #include <libtrx/game/lara/control.h>
-
-void Lara_HandleAboveWater(ITEM *item, COLL_INFO *coll);
-void Lara_HandleUnderwater(ITEM *item, COLL_INFO *coll);
-void Lara_HandleSurface(ITEM *item, COLL_INFO *coll);
