+++ src/tr2/game/lara/control.h
@@ -5,19 +5,8 @@
 
 #include <libtrx/game/lara/common.h>
 
-void Lara_HandleAboveWater(ITEM *item, COLL_INFO *coll);
-
-void Lara_HandleSurface(ITEM *item, COLL_INFO *coll);
-
-void Lara_HandleUnderwater(ITEM *item, COLL_INFO *coll);
-
-void Lara_Control(int16_t item_num);
 void Lara_ControlExtra(int16_t item_num);
-
 void Lara_UseItem(GAME_OBJECT_ID obj_id);
-
 void Lara_InitialiseLoad(int16_t item_num);
-
 void Lara_Initialise(const GF_LEVEL *level);
-
 void Lara_InitialiseInventory(const GF_LEVEL *level);
